refactor: change qualifiers to use cloud instead of standard.

+++ version-service/src/main/kotlin/com/jtm/version/core/domain/model/FolderInfo.kt
@@ -1,3 +1,3 @@
 package com.jtm.version.core.domain.model
 
-data class FolderInfo(val name: String, val files: Int)
+data class FolderInfo(val name: String, val files: Long)
